Updated correct jupyter file uploaded.
The original file had mistake on steps 5 and 6. the variable name was defined as 'acc', however the correct variable name is 'accuracy' as that is what was asked for the users to write in the instructions given in step 5 as "<addMetric> with 'accuracy' ". This is something which is difficult to remove for a beginer.